fazendo requisicoes no front

+++ server/front/pages/game/script.js
@@ -1,27 +1,35 @@
-import { level, setGetId } from "./requisitions.js";
+import { level, setGetAccount, addPoint } from "./requisitions.js";
+let pieces;
+$(function(){
+    try{
+        const session = JSON.parse(sessionStorage.getItem('account'))
+        if(session === null) throw new Error('error');
+        setGetAccount(session);
+        sessionStorage.clear();
+        requisitions();
+    }
+    catch(e){
+        sessionStorage.clear();
+        window.location.href = "http://localhost:8000/"
+    }
+});
 
-try{
-    const session = JSON.parse(sessionStorage.getItem('account'));
-    setGetId(session.id);
-    sessionStorage.clear();
-}
-catch(e){
-    sessionStorage.clear();
-    window.location.href = "http://localhost:8000/"
+async function requisitions(){
+    const lvl = await level(await setGetAccount().level);  
+    printStage(lvl.stage);
+    pieces = lvl.pieces
+    printPieces();
 }
 
 let revertPiece;
 let dragging = 0;
 const audioDrag = new Audio("https://assets.mixkit.co/sfx/preview/mixkit-money-bag-drop-1989.mp3");
-
 const audioDrop = new Audio("https://assets.mixkit.co/sfx/preview/mixkit-small-wood-plank-pile-drop-3141.mp3");
 
-$(function () {
-    printStage();
-    printPieces();
-    
+//Set events every elements
+function setEvents(){    
     $('.piece').draggable({
-        snap: '.stage-valid',
+        snap: '.stage-valid' ,
         handle: '.piece-color',
         cursor: "grabbing",
         stack: ".piece",
@@ -30,7 +38,7 @@ $(function () {
             $(audioDrag)[0].play();
         },
         stop: function( event, ui) {
-            points++;            
+            console.log(addPoint());            
             validation(event);
         },
         revert: function(){
@@ -40,39 +48,18 @@ $(function () {
                 return false
             }
         },
-        drag: function(event, ui){
-            validation(event);
-        },
         addClasses: false
         
     });
 
     $('.piece-color').dblclick(rotate)
-});
+}
 
-let points = 0
-jQuery.fn.single_double_click = function(single_click_callback, double_click_callback, timeout) {
-    return this.each(function(){
-      var clicks = 0, self = this;
-      jQuery(this).click(function(event){
-        clicks++;
-        if (clicks == 1) {
-          setTimeout(function(){
-            if(clicks == 1) {
-              single_click_callback.call(self, event);
-            } else {
-              double_click_callback.call(self, event);
-            }
-            clicks = 0;
-          }, timeout || 300 );
-        }
-      });
-    });
-  }
-
-
-function printStage() {
-    const arr = [[0, 1, 1, 1, 0], [1, 1, 1, 1, 1], [1, 1, 1, 1, 1], [0, 1, 1, 1, 0]];
+//--------------------Print Stage
+function printStage(arr) {
+    $('section').html('');
+    $('section').append('<div id="all"></div>')
+    //const arr = [[0, 1, 1, 1, 0], [1, 1, 1, 1, 1], [1, 1, 1, 1, 1], [0, 1, 1, 1, 0]];
     $('section').append('<div class="stage"></div>');
     //forEach piece of the stage
     arr.forEach(elem => {
@@ -83,23 +70,8 @@ function printStage() {
         }
     });
 }
-const exFetch = {"level": 1, "difficulty": "Fácil", "stage": [ [1,0,0,0,1],[0,0,0,0,0],[0,0,0,0,0],[0,0,0,0,0],[1,0,0,0,1] ], "pieces": [ [ [0,1],[1,1] ], [ [1,1], [1,1] ], [ [1,1,1] ], [ [1,1,1] ], [ [1,1,1],[0,0,1] ], [ [0,0,1],[1,1,1] ] ]};
 
-const pieces = [
-    {id:'piece1', shape: [[1,1,1]], deg: 0},
-    {id:'piece2', shape: [[1,1,1]], deg: 0},
-    {id:'piece3', shape: [[1,1,1]], deg: 0},
-    {id:'piece4', shape: [[1,1,1]], deg: 0},
-    {id:'piece5', shape: [[1,1],[1,1]], deg: 0},
-    {id:'piece6', shape: [[1,0],[1,1],[0,1]], deg: 0},
-    {id:'piece7', shape: [[1,0],[1,1],[0,1]], deg: 0},
-    {id:'piece8', shape: [[1,1,1]], deg: 0},
-    {id:'piece9', shape: [[1,1,1]], deg: 0},
-    {id:'piece10', shape: [[1,1,1]], deg: 0},
-    {id:'piece11', shape: [[1,1],[1,1]], deg: 0}
-];
-
-
+//------------------- Print Pieces
 function printPieces(){
     //forEach in the piece
     pieces.forEach(elem => {
@@ -115,30 +87,32 @@ function printPieces(){
             })
         })
     });
+    setEvents();
 }
 
+// ------------------Rotate
 function rotate(e){
-    points++;
+    console.log(addPoint());
     const _e = e.target.parentNode.offsetParent;   
     const _eId = _e.id.replace(/\D/g,'');
     const piece = pieces[_eId-1];
 
-    $(_e).addClass('.piece-rotate');
     $(_e).css('transition', '500ms');
     piece.deg += 90;    
+
     _e.style.transform = `rotate(${piece.deg}deg)`;
-    var parent = e.target.parentNode;
-    var grandParent = parent.parentNode;
+    const parent = e.target.parentNode;
+    const grandParent = parent.parentNode;
     $(grandParent).draggable("disable");
     setTimeout(() => {
         $(_e).css('transition', '0ms');
         $(_e).removeClass('.piece-rotate');
         $(grandParent).draggable("enable");
     }, 500);
-    
 }
 
-//Validation pieces on stage
+
+// ------------------Validation pieces on stage
 function validation(e){
     revertPiece = 0;
     const allPieces = document.getElementsByClassName('piece-color');
@@ -171,8 +145,12 @@ function validation(e){
     //Here is the validation
     if(validOrInvalid.every(elem => elem)){
         $('.piece-color').css('background', 'green');
-    }else{
-        $('.piece-color').css('background', 'yellow');
+        outraFase();
     }
     e.target.offsetParent.style.zIndex = 0;
+}
+
+function outraFase(){
+    console.log(setGetAccount('add'));
+    requisitions();
 }
